tests: platforminfo: skip DBus test if hostname service isn't available
The dbus test is dependent on the org.freedesktop.hostname1 service
being present. If that isn't there, we really can't determine whether or
not the PlatformInfo code does what it's supposed to do.

So try to determine whether or not it makes sense to execute this test
in the first place.